recover my_assistant inst. and migrate main.py contents
