+++ AWShealthCoach/make_assistant.py
@@ -2,6 +2,16 @@ from litellm import completion
 import os
 from typing import List, Dict
 
+from prompts import *
+from make_assistant import *
+
+from context_management.compression import ContextCompressor
+from memory import Memory
+
+from langchain_community.document_loaders import PyPDFLoader
+from langchain.schema import Document
+
+
 # from dotenv import load_dotenv
 # load_dotenv()
 
@@ -40,14 +50,57 @@ def generate_response(messages: List[Dict[str, str]], model: str = "gpt-4-0125-p
 #   purpose='assistants'
 # )
 
-# my_assistant = client.beta.assistants.create(
-#     instructions=instructions_str,
-#     name="Guide v3",
-#     tools=[{"type": "retrieval"}],
-#     model="gpt-4-turbo-preview",
-#     file_ids=[file.id]
-# )
-# print(my_assistant)
+def ingest_documents(file_paths: List[str]) -> List[Document]:
+    documents_list = []
+    for file_path in file_paths:
+        if file_path.endswith('.pdf'):
+            loader = PyPDFLoader(file_path)
+            pages = loader.load_and_split()
+            documents_list.extend(pages)
+        else:
+            raise ValueError(f"File {file_path} must be a PDF.")
+    return documents_list
+
+def pratik_patel_response(user_message: str, embedding_provider: str = "openai") -> str:
+
+    instruct_prompt = pratik_patel_instruct()
+
+    #TODO probably need to parameterize file inputs
+    documents = ingest_documents(['text_corpus/survey_questions_and_answers.pdf'])
+    memory = Memory(embedding_provider)
+    context = ContextCompressor(documents=documents, embeddings=memory.get_embeddings())
+
+    #TODO adjust max_results arg as needed
+    relevant_context = context.get_context(user_message, max_results=5)
+
+    message = [
+        {"role": "user", "content": relevant_context},
+        {"role": "user", "content": user_message},
+    ]
+    return generate_response(messages=message)
+
+my_assistant = client.beta.assistants.create(
+    instructions=instructions_str,
+    name="Guide v3",
+    tools=[{
+        "type": "function",
+      "function": {
+      "name": "pratik_patel_response",
+      "description": "Generate an answer as Pratik Patel given a user's input question.",
+      "parameters": {
+        "type": "object",
+        "properties": {
+          "user_message": {"type": "string", "description": "The user's input question."},
+          "embedding_provider": {"type": "string", "enum": ["openai"], "default": "openai", "description": "The embedding provider to use."}
+        },
+        "required": ["user_message"]
+        }
+      }
+    }],
+    model="gpt-4-turbo-preview",
+    file_ids=[file.id]
+)
+print(my_assistant)
 
 
 # initialize_app()
